==> notebooks/01_load_raw_data.ipynb <==
from pathlib import Path
import requests

def download_raw_data_one_file(year: int, month: int) -> Path:
    ''''''
    URL = f'https://d37ci6vzurychx.cloudfront.net/trip-data/yellow_tripdata_{year}-{month:02d}.parquet'
    response = requests.get(URL)

    if response.status_code == 200:
        path = f'../data/raw/rides_{year}-{month:02d}.parquet'
        open(path, "wb").write(response.content)
        return path
    else:
        raise Exception(f'{URL} is not available')

download_raw_data_one_file(2022, 1)

import pandas as pd

rides = pd.read_parquet('../data/raw/rides_2022-01.parquet')

rides.head(20)

rides = rides[['tpep_pickup_datetime', 'PULocationID']]

rides.rename(columns={'tpep_pickup_datetime': 'pickup_datetime', 'PULocationID': 'pickup_location_id'}, inplace=True)

rides

# validate data
rides['pickup_datetime'].describe()

def validate_raw_data(rides: pd.DataFrame, year: int, month: int) -> pd.DataFrame:
    '''Removes rows with pickup_datetime outside of the given year and month'''
    
    rides = rides.copy()
    rides['pickup_datetime'] = pd.to_datetime(rides['pickup_datetime'])
    rides = rides[(rides['pickup_datetime'].dt.year == year) & (rides['pickup_datetime'].dt.month == month)]

    return rides

rides = validate_raw_data(rides=rides, year=2022, month=1)
rides

rides['pickup_datetime'].describe()

rides.to_parquet('../data/transformed/validated_rides_2022_01.parquet')
