==> notebooks/02_transform_raw_data_to_ts_data.ipynb <==
import pandas as pd

rides = pd.read_parquet('../data/transformed/validated_rides_2022_01.parquet')

rides.head()

rides['pickup_hour'] = rides['pickup_datetime'].dt.floor('H')
rides

agg_rides = rides.groupby(['pickup_hour', 'pickup_location_id']).size().reset_index()
agg_rides.rename(columns={0: 'rides'}, inplace=True)
agg_rides

from tqdm import tqdm

def add_missing_dates(agg_rides: pd.DataFrame) -> pd.DataFrame:
    '''Takes an aggregated rides dataframe and adds
    missing dates to the dataframe with a frequency of hourly.
    The function returns the dataframe with the added missing dates.
    '''

    
    location_ids = agg_rides['pickup_location_id'].unique()
    full_range = pd.date_range(agg_rides['pickup_hour'].min(), agg_rides['pickup_hour'].max(), freq='H')

    output = pd.DataFrame()

    for location_id in tqdm(location_ids):

        agg_rides_i = agg_rides[agg_rides['pickup_location_id'] == location_id][['pickup_hour', 'rides']]

        agg_rides_i.set_index('pickup_hour', inplace=True)
        agg_rides_i.index = pd.DatetimeIndex(agg_rides_i.index)
        agg_rides_i = agg_rides_i.reindex(full_range, fill_value=0)

        agg_rides_i['pickup_location_id'] = location_id

        output = pd.concat([output, agg_rides_i])

    output = output.reset_index().rename(columns={'index': 'pickup_hour'})

    return output

agg_rides_all_dates = add_missing_dates(agg_rides)
agg_rides_all_dates

agg_rides_all_dates.to_parquet('../data/transformed/ts_data_2022_01.parquet')

from typing import List, Optional
import plotly.express as px
import pandas as pd

def plot_rides(rides: pd.DataFrame, locations: Optional[List[int]] = None):
    '''Takes a rides dataframe and plots the rides for each location.
    '''

    if locations is None:
        locations = rides['pickup_location_id'].unique()

    for location_id in locations:
        fig = px.line(rides[rides['pickup_location_id'] == location_id], x='pickup_hour', y='rides')
        fig.show()

plot_rides(agg_rides_all_dates, locations=[43])
