==> notebooks/03_transform_ts_data_to_features_and_target.ipynb <==
import pandas as pd
from tqdm import tqdm

ts_data = pd.read_parquet('../data/transformed/ts_data_2022_01.parquet')
ts_data

ts_data_one_location = ts_data[ts_data['pickup_location_id'] == 43].reset_index(drop=True)
ts_data_one_location.head(25)

def create_ts_dataset(data: pd.DataFrame, n_features: int, step_size: int) -> pd.DataFrame:
    """
    Create a dataset with n_features and a target column based on step_size.
    """

    import warnings

    warnings.simplefilter(action='ignore', category=pd.errors.PerformanceWarning)

    df_features = pd.DataFrame()
    df_target = pd.DataFrame()
    for location_id in tqdm(data['pickup_location_id'].unique()):
        data_i = data[data['pickup_location_id'] == location_id].reset_index(drop=True)
    
        df_features_i = pd.DataFrame()
        df_target_i = pd.DataFrame()
        # take the last n_features rows and add them as features
        for i in range(n_features):
            df_features_i['pickup_hour'] = data_i['pickup_hour'].shift(-n_features)
            df_features_i['pickup_location_id'] = data_i['pickup_location_id']
            df_features_i[f'rides_previous_{n_features-i}_hour'] = data_i['rides'].shift(-i)

            # take the next row after the last n_features rows and add it as target
            df_target_i['target_rides_next_hour'] = data_i['rides'].shift(-n_features)
    
        df_features = pd.concat([df_features, df_features_i])
        df_target = pd.concat([df_target, df_target_i])

    # select rows based on step_size
    index = list(range(0, len(df_features), step_size))
    df_features = df_features.iloc[index]
    df_target = df_target.iloc[index]
    
    return df_features.dropna().reset_index(drop=True), df_target['target_rides_next_hour'].dropna().reset_index(drop=True)

n_features = 24*7*1 # 1 week
step_size = 24

features, target = create_ts_dataset(ts_data, n_features, step_size)

features

target
