==> notebooks/04_transformed_raw_data_to_features_and_targets.ipynb <==
# autorelad modules
%load_ext autoreload
%autoreload 2

import os
import numpy as np

# change working directory to project root
os.chdir(os.path.dirname(os.getcwd()))

from src.data import load_raw_data

rides = load_raw_data(2022)
rides

from src.data import transform_raw_data_into_ts_data

ts_data = transform_raw_data_into_ts_data(rides)
ts_data

from src.data import create_ts_dataset

features, targets = create_ts_dataset(ts_data, n_features=24*28*1, step_size=24)

print(f'{features.shape=}')
print(f'{targets.shape=}')

# Ensure each location and date has the same number of rows
assert np.sum(features.groupby(['pickup_location_id', features['pickup_hour'].dt.date]).count().reset_index(drop=True).values) == len(features) * (len(features.columns) - 1), 'Each location ID & Date does not have the same number of rows.'

tabular_data = features
tabular_data['target_rides_next_hour'] = targets

from src.paths import TRANSFORMED_DATA_DIR
tabular_data.to_parquet(TRANSFORMED_DATA_DIR / 'tabular_data.parquet')

tabular_data
