==> notebooks/06_baseline_models.ipynb <==
import pandas as pd

# set current working directory to project root
import os
os.chdir(os.path.dirname(os.getcwd()))
from src.paths import TRANSFORMED_DATA_DIR

df = pd.read_parquet(TRANSFORMED_DATA_DIR / 'tabular_data.parquet')
df

from datetime import datetime
from src.data_split import train_test_split

X_train, y_train, X_test, y_test = train_test_split(df, datetime(2022, 6, 1, 0, 0, 0), 'target_rides_next_hour')

print(f'{X_train.shape=}')
print(f'{y_train.shape=}')
print(f'{X_test.shape=}')
print(f'{y_test.shape=}')


# Baseline Model 1
- Actual demand observed in the previous hour

import numpy as np

# LOCF baseline model
class BaselineModelLOCF:
    '''Prediction = actual demand observed in the previous hour'''
    
    def fit(self, X_train: pd.DataFrame, y_train: pd.Series):
        pass
        
    def predict(self, X_test: pd.DataFrame) -> np.array:
        return X_test[f'rides_previous_1_hour']

model = BaselineModelLOCF()
predictions = model.predict(X_test)
predictions

from sklearn.metrics import mean_absolute_error

test_mae = mean_absolute_error(y_test, predictions)
print(f'{test_mae=:.4f}')

# Baseline Model 2
- Actual demand observed at t-7 days

import pandas as pd
import numpy as np

# previous week baseline model
class BaselineModelPreviousWeek:
    '''Prediction = actual demand observed at t-7 days'''
    
    def fit(self, X_train: pd.DataFrame, y_train: pd.Series):
        pass
        
    def predict(self, X_test: pd.DataFrame) -> np.array:
        return X_test[f'rides_previous_{7*24}_hour']

model = BaselineModelPreviousWeek()
predictions = model.predict(X_test)
predictions

test_mae = mean_absolute_error(y_test, predictions)
print(f'{test_mae=:.4f}')

# Baseline Model 3
- Actual demand observed at t-7 days, t-14 days, t-21 days, t-28

import pandas as pd
import numpy as np

# average of previous 4 weeks baseline model
class BaselineModelPrevious4Weeks:
    '''Prediction = actual demand observed at t-7 days, t-14 days, t-21 days, t-28'''
    
    def fit(self, X_train: pd.DataFrame, y_train: pd.Series):
        pass
        
    def predict(self, X_test: pd.DataFrame) -> np.array:
        return (
            X_test[f'rides_previous_{7*24}_hour'] + \
            X_test[f'rides_previous_{14*24}_hour'] + \
            X_test[f'rides_previous_{21*24}_hour'] + \
            X_test[f'rides_previous_{28*24}_hour']
            ) / 4

model = BaselineModelPrevious4Weeks()
predictions = model.predict(X_test)
predictions

test_mae = mean_absolute_error(y_test, predictions)
print(f'{test_mae=:.4f}')
