==> notebooks/09_lightgbm_model_with_feature_engineering.ipynb <==
# autoreload
%load_ext autoreload
%autoreload 2

from datetime import datetime
import requests
import pandas as pd
import lightgbm as lgb

from sklearn.preprocessing import FunctionTransformer
from sklearn.base import BaseEstimator, TransformerMixin
from sklearn.pipeline import make_pipeline
from sklearn.metrics import mean_absolute_error


# set current working directory to project root
import os
os.chdir(os.path.dirname(os.getcwd()))
from src.paths import TRANSFORMED_DATA_DIR, RAW_DATA_DIR
from src.data_split import train_test_split
from src.data import download_and_load_nyc_taxi_zone_data

# load dataset
df = pd.read_parquet(TRANSFORMED_DATA_DIR / 'tabular_data.parquet')
df

# train test split
X_train, y_train, X_test, y_test = train_test_split(df, datetime(2022, 6, 1, 0, 0, 0), 'target_rides_next_hour')

print(f'{X_train.shape=}')
print(f'{y_train.shape=}')
print(f'{X_test.shape=}')
print(f'{y_test.shape=}')

# function that averages rides from previous 7, 14, 21, 28 days
def average_rides_last_4_weeks(X: pd.DataFrame) -> pd.DataFrame:
    '''Adds a feature column calculated by averaging rides from
    - previous 7 days
    - previous 14 days
    - previous 21 days
    - previous 28 days
    '''

    X_ = X.copy()
    X_['average_rides_last_4_weeks'] = X_[[f'rides_previous_{7*24}_hour', f'rides_previous_{14*24}_hour', f'rides_previous_{21*24}_hour', f'rides_previous_{28*24}_hour']].mean(axis=1)

    return X_

# convert function to sklearn transformer
add_feature_average_rides_last_4_weeks = FunctionTransformer(average_rides_last_4_weeks, validate=False)

def ex_tract_temporal_features(X: pd.DataFrame) -> pd.DataFrame:
    '''Ex_tracts temporal features from the datetime index_'''

    X_ = X.copy()
    X_['hour'] = X_.pickup_hour.dt.hour
    X_['day'] = X_.pickup_hour.dt.day
    X_['month'] = X_.pickup_hour.dt.month
    X_['weekday'] = X_.pickup_hour.dt.weekday
    X_['weekend'] = X_['weekday'].isin([5, 6]).astype(int)
    X_.drop('pickup_hour', axis=1, inplace=True)
    
    return X_

# convert function to sklearn transformer
add_temporal_features = FunctionTransformer(ex_tract_temporal_features, validate=False)

# download and load latitude and longitude data
nyc_zone_data = download_and_load_nyc_taxi_zone_data()
nyc_zone_data

# function to extract latitude & longitude from new york city location id data
def extract_lat_lon(df: pd.DataFrame) -> pd.DataFrame:
    '''Extracts latitude and longitude from the_geom column'''
    
    df_ = df.copy()
    # extract latitude and longitude data from the_geom column
    df_['the_geom'] = df_['the_geom'].str.replace('MULTIPOLYGON \\(\\(\\(', '', regex=True).str.replace('\\)\\)\\)', '', regex=True).str.replace(',', '', regex=True).str.replace(' ', ',', regex=True).str.replace('\\)\\)', '', regex=True).str.replace('\\(\\(', '', regex=True)

    # convert string to float
    df_['the_geom'] = df_['the_geom'].apply(lambda x: [float(i) for i in x.split(',')])

    # convert list of floats to list of tuples
    df_['the_geom'] = df_['the_geom'].apply(lambda x: list(zip(x[::2], x[1::2])))

    # take average of tuples to get center of zone
    df_['the_geom'] = df_['the_geom'].apply(lambda x: (sum([i[0] for i in x])/len(x), sum([i[1] for i in x])/len(x)))

    df_['latitude'] = df_['the_geom'].apply(lambda x: x[0])
    df_['longitude'] = df_['the_geom'].apply(lambda x: x[1])
    
    return df_[['LocationID', 'latitude', 'longitude']]

nyc_zone_data = extract_lat_lon(nyc_zone_data)

# convert function to sklearn transformer
class add_latitude_and_longitude_features(BaseEstimator, TransformerMixin):
    """add latitude and longitude features to the dataframe"""

    def __init__(self, df_lat_lon: pd.DataFrame):
        self.df_lat_lon = df_lat_lon

    def fit(self, X, y=None):
        return self
    
    def transform(self, X, y=None):
        
        X_ = X.copy()
        
        # merge latitude and longitude data with X
        X_ = X_.merge(self.df_lat_lon, how='left', left_on='pickup_location_id', right_on='LocationID')

        # drop LocationID column
        X_.drop('LocationID', axis=1, inplace=True)

        # rename columns
        X_.rename(columns={'latitude': 'pickup_latitude', 'longitude': 'pickup_longitude'}, inplace=True)

        return X_

# Create pipeline

pipeline = make_pipeline(
    add_feature_average_rides_last_4_weeks,
    add_temporal_features,
    add_latitude_and_longitude_features(nyc_zone_data),
    lgb.LGBMRegressor()
)
pipeline.fit(X_train, y_train)

predictions = pipeline.predict(X_test)


test_mae = mean_absolute_error(y_test, predictions)
print(f'{test_mae=:.4f}')
